pyproject.toml
[project]
name = "narratix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "flask>=3.1.0",
    "google-generativeai>=0.8.4",
    "groq>=0.22.0",
    "nltk>=3.9.1",
    "requests>=2.32.3",
    "textblob>=0.19.0",
]
